Move Bandit scripts from Entity/Enemy/* to Entity/Enemy/Bandit/*
