+++ bci-game/Assets/Scripts/Entity/Enemy/Bandit.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: cf227a5f30594de0beaf42fdf3aea670
+timeCreated: 1704406217
